Add new features to Rare CLI
The script now lists all available pods in the namespace, with the
ability to select a pod. This should make scaling the environment
easier. There is also a new feature to destroy and re-create a pod, a
verbose argument, and a shell argument.

Old usage: ./rare.sh qlora
New usage: ./rare.sh shell qlora or ./rare.sh sh qlor